moved js to folders
