+++ src/catalog.js
@@ -1,7 +1,7 @@
 import './javascript/loader';
-import './javascript/header';
-import './javascript/hero';
-import './javascript/catalog-page';
-import './javascript/modal-trailer';
+import './javascript/header/header';
+import './javascript/hero/hero';
+import './javascript/catalog/catalog-page';
+import './javascript/modal/modal-trailer';
 import './javascript/global-listener';
 import './javascript/scroll-up';
