+++ src/index.js
@@ -1,10 +1,10 @@
 import './javascript/loader';
 import './javascript/stars-rating';
-import './javascript/header';
-import './javascript/hero';
-import './javascript/weekly-trends-main';
-import './upcoming/upcoming';
-import './javascript/modal-trailer';
+import './javascript/header/header';
+import './javascript/hero/hero';
+import './javascript/weekly-trends/weekly-trends-main';
+import './javascript/upcoming/upcoming';
+import './javascript/modal/modal-trailer';
 import './javascript/global-listener';
 import './javascript/scroll-up';
 // console.log('object');
