+++ src/javascript/global-listener.js
@@ -1,14 +1,13 @@
-import { openModalAboutFilm } from './movieModal';
+import { openModalAboutFilm } from './modal/movieModal';
 import { handleFilm } from './library/library';
 // import { playTrailer } from './modal-trailer';
 
-window.addEventListener('click', (e) => {
+window.addEventListener('click', e => {
   const heroDetailsButton = document.querySelector('.css-bnt-info');
   const watchTrailerButton = document.querySelector('.watch-trailer-button');
   // const addBtn = document.querySelector('[data-add]');
   // const removeBtn = document.querySelector('[data-remove]');
   const upcomingBtn = document.querySelector('.btn');
-
 
   const id = e.target.dataset.id;
 
@@ -22,8 +21,8 @@ window.addEventListener('click', (e) => {
     //   break;
 
     case upcomingBtn:
-    // case removeBtn:
-      handleFilm(e)
+      // case removeBtn:
+      handleFilm(e);
       break;
 
     case watchTrailerButton:
