+++ src/javascript/hero/hero.js
@@ -1,5 +1,5 @@
-import starsRating from './stars-rating';
-import { getMovieId } from './modal-trailer';
+import starsRating from '../stars-rating';
+import { getMovieId } from '../modal/modal-trailer';
 
 const API_KEY = '58fde9f9a3392c3dbee86a1f2142354e';
 const RANDOM_NUMBER = Math.floor(Math.random() * (19 - 0 + 1)) + 0;
@@ -17,7 +17,6 @@ window.addEventListener('load', heroInfoShow);
 
 function heroInfoShow() {
   getFilmInfo().then(({ data }) => {
-
     // Змінні усієї інформації фільму
     const filmInfo = data.results[RANDOM_NUMBER];
     const filmPicturePath = filmInfo.backdrop_path;
