+++ src/javascript/modal/movieDetailProvider.js
