+++ src/javascript/upcoming/helpers.js
