+++ src/javascript/upcoming/img-handler.js
