+++ src/javascript/weekly-trends/weekly-trends-genres.js
