+++ src/library.js
@@ -1,4 +1,3 @@
 import './javascript/loader';
-import './javascript/header';
 import './javascript/scroll-up';
 import './javascript/library/library';
